refactor: Minor cleanup of test project.
